els-ASAP auto commit file/tag dbus-1.10.24-12.el7.src.rpm

+++ src/tools/dbus-launch.c
@@ -38,10 +38,18 @@
 #include <sys/select.h>
 #include <time.h>
 
+#include <dbus/dbus.h>
+#include "dbus/dbus-internals.h"
+
 #ifdef DBUS_BUILD_X11
 #include <X11/Xlib.h>
 extern Display *xdisplay;
 #endif
+
+#include "dbus/dbus-internals.h"
+#include "dbus/dbus-sysdeps-unix.h"
+
+#include "tool-common.h"
 
 /* PROCESSES
  *
@@ -102,56 +110,33 @@ save_machine_uuid (const char *uuid_arg)
       exit (1);
     }
 
-  machine_uuid = xstrdup (uuid_arg);
+  machine_uuid = _dbus_strdup (uuid_arg);
 }
 
-#define UUID_MAXLEN 40
+#ifdef DBUS_BUILD_X11
 /* Read the machine uuid from file if needed. Returns TRUE if machine_uuid is
  * set after this function */
 static int
 read_machine_uuid_if_needed (void)
 {
-  FILE *f;
-  char uuid[UUID_MAXLEN];
-  size_t len;
-  int ret = FALSE;
-
   if (machine_uuid != NULL)
     return TRUE;
 
-  f = fopen (DBUS_MACHINE_UUID_FILE, "r");
-  if (f == NULL)
+  machine_uuid = dbus_get_local_machine_id ();
+
+  if (machine_uuid == NULL)
     return FALSE;
 
-  if (fgets (uuid, UUID_MAXLEN, f) == NULL)
-    goto out;
-
-  len = strlen (uuid);
-  if (len < 32)
-    goto out;
-
-  /* rstrip the read uuid */
-  while (len > 31 && isspace(uuid[len - 1]))
-    len--;
-
-  if (len != 32)
-    goto out;
-
-  uuid[len] = '\0';
-  machine_uuid = xstrdup (uuid);
   verbose ("UID: %s\n", machine_uuid);
-  ret = TRUE;
-
-out:
-  fclose(f);
-  return ret;
+  return TRUE;
 }
-
+#endif /* DBUS_BUILD_X11 */
 
 void
 verbose (const char *format,
          ...)
 {
+#ifdef DBUS_ENABLE_VERBOSE_MODE
   va_list args;
   static int verbose = TRUE;
   static int verbose_initted = FALSE;
@@ -176,12 +161,16 @@ verbose (const char *format,
   va_start (args, format);
   vfprintf (stderr, format, args);
   va_end (args);
+#endif /* DBUS_ENABLE_VERBOSE_MODE */
 }
 
 static void
 usage (int ecode)
 {
-  fprintf (stderr, "dbus-launch [--version] [--help] [--sh-syntax] [--csh-syntax] [--auto-syntax] [--exit-with-session]\n");
+  fprintf (stderr, "dbus-launch [--version] [--help] [--sh-syntax]"
+           " [--csh-syntax] [--auto-syntax] [--binary-syntax] [--close-stderr]"
+           " [--exit-with-session] [--autolaunch=MACHINEID]"
+           " [--config-file=FILENAME] [PROGRAM] [ARGS...]\n");
   exit (ecode);
 }
 
@@ -214,6 +203,26 @@ xstrdup (const char *str)
   memcpy (copy, str, len + 1);
   
   return copy;
+}
+
+static char *
+concat2 (const char *a,
+    const char *b)
+{
+  size_t la, lb;
+  char *ret;
+
+  la = strlen (a);
+  lb = strlen (b);
+
+  ret = malloc (la + lb + 1);
+
+  if (ret == NULL)
+    return NULL;
+
+  memcpy (ret, a, la);
+  memcpy (ret + la, b, lb + 1);
+  return ret;
 }
 
 typedef enum
@@ -380,6 +389,9 @@ static pid_t bus_pid_to_kill = -1;
 static void
 kill_bus(void)
 {
+  if (bus_pid_to_kill <= 0)
+    return;
+
   verbose ("Killing message bus and exiting babysitter\n");
   kill (bus_pid_to_kill, SIGTERM);
   sleep (3);
@@ -414,7 +426,8 @@ print_variables (const char *bus_address, pid_t bus_pid, long bus_wid,
   else if (c_shell_syntax)
     {
       printf ("setenv DBUS_SESSION_BUS_ADDRESS '%s';\n", bus_address);	
-      printf ("set DBUS_SESSION_BUS_PID=%ld;\n", (long) bus_pid);
+      if (bus_pid)
+        printf ("set DBUS_SESSION_BUS_PID=%ld;\n", (long) bus_pid);
       if (bus_wid)
         printf ("set DBUS_SESSION_BUS_WINDOWID=%ld;\n", (long) bus_wid);
       fflush (stdout);
@@ -423,7 +436,8 @@ print_variables (const char *bus_address, pid_t bus_pid, long bus_wid,
     {
       printf ("DBUS_SESSION_BUS_ADDRESS='%s';\n", bus_address);
       printf ("export DBUS_SESSION_BUS_ADDRESS;\n");
-      printf ("DBUS_SESSION_BUS_PID=%ld;\n", (long) bus_pid);
+      if (bus_pid)
+        printf ("DBUS_SESSION_BUS_PID=%ld;\n", (long) bus_pid);
       if (bus_wid)
         printf ("DBUS_SESSION_BUS_WINDOWID=%ld;\n", (long) bus_wid);
       fflush (stdout);
@@ -431,7 +445,8 @@ print_variables (const char *bus_address, pid_t bus_pid, long bus_wid,
   else
     {
       printf ("DBUS_SESSION_BUS_ADDRESS=%s\n", bus_address);
-      printf ("DBUS_SESSION_BUS_PID=%ld\n", (long) bus_pid);
+      if (bus_pid)
+        printf ("DBUS_SESSION_BUS_PID=%ld\n", (long) bus_pid);
       if (bus_wid)
 	printf ("DBUS_SESSION_BUS_WINDOWID=%ld\n", (long) bus_wid);
       fflush (stdout);
@@ -445,9 +460,7 @@ signal_handler (int sig)
 {
   switch (sig)
     {
-#ifdef SIGHUP
     case SIGHUP:
-#endif
     case SIGINT:
     case SIGTERM:
       got_sighup = TRUE;
@@ -509,7 +522,7 @@ kill_bus_when_session_ends (void)
   if (tty_fd < 0 && x_fd < 0)
     {
       fprintf (stderr, "No terminal on standard input and no X display; cannot attach message bus to session lifetime\n");
-      exit (1);
+      kill_bus_and_exit (1);
     }
   
   while (TRUE)
@@ -755,31 +768,35 @@ pass_info (const char *runprog, const char *bus_address, pid_t bus_pid,
       if (envvar == NULL || args == NULL)
         goto oom;
 
-     args[0] = xstrdup (runprog);
+      args[0] = xstrdup (runprog);
       if (!args[0])
         goto oom;
-     for (i = 1; i <= (argc-remaining_args); i++)
-      {
-        size_t len = strlen (argv[remaining_args+i-1])+1;
-        args[i] = malloc (len);
-        if (!args[i])
-          goto oom;
-        strncpy (args[i], argv[remaining_args+i-1], len);
-       }
-     args[i] = NULL;
+      for (i = 1; i <= (argc-remaining_args); i++)
+        {
+          size_t len = strlen (argv[remaining_args+i-1])+1;
+          args[i] = malloc (len);
+          if (!args[i])
+	    {
+              while (i > 1)
+                free (args[--i]);
+              goto oom;
+	    }
+          strncpy (args[i], argv[remaining_args+i-1], len);
+        }
+      args[i] = NULL;
 
-     strcpy (envvar, "DBUS_SESSION_BUS_ADDRESS=");
-     strcat (envvar, bus_address);
-     putenv (envvar);
+      strcpy (envvar, "DBUS_SESSION_BUS_ADDRESS=");
+      strcat (envvar, bus_address);
+      putenv (envvar);
 
-     execvp (runprog, args);
-     fprintf (stderr, "Couldn't exec %s: %s\n", runprog, strerror (errno));
-     exit (1);
+      execvp (runprog, args);
+      fprintf (stderr, "Couldn't exec %s: %s\n", runprog, strerror (errno));
+      exit (1);
     }
    else
     {
       print_variables (bus_address, bus_pid, bus_wid, c_shell_syntax,
-         bourne_shell_syntax, binary_syntax);
+          bourne_shell_syntax, binary_syntax);
     }
   verbose ("dbus-launch exiting\n");
 
@@ -823,10 +840,27 @@ main (int argc, char **argv)
   int bus_pid_to_babysitter_pipe[2];
   int bus_address_to_launcher_pipe[2];
   char *config_file;
-  
+  dbus_bool_t user_bus_supported = FALSE;
+  DBusString user_bus;
+  const char *error_str;
+
   exit_with_session = FALSE;
   config_file = NULL;
-  
+
+  /* Ensure that the first three fds are open, to ensure that when we
+   * create other file descriptors (for example for epoll, inotify or
+   * a socket), they never get assigned as fd 0, 1 or 2. If they were,
+   * which could happen if our caller had (incorrectly) closed those
+   * standard fds, then we'd start dbus-daemon with those fds closed,
+   * which is unexpected and could cause it to misbehave. */
+  if (!_dbus_ensure_standard_fds (0, &error_str))
+    {
+      fprintf (stderr,
+               "dbus-launch: fatal error setting up standard fds: %s: %s\n",
+               error_str, _dbus_strerror (errno));
+      return 1;
+    }
+
   prev_arg = NULL;
   i = 1;
   while (i < argc)
@@ -976,11 +1010,43 @@ main (int argc, char **argv)
       char *address;
       pid_t pid;
       long wid;
+      DBusError error = DBUS_ERROR_INIT;
       
       if (get_machine_uuid () == NULL)
         {
           fprintf (stderr, "Machine UUID not provided as arg to --autolaunch\n");
           exit (1);
+        }
+
+      if (!_dbus_string_init (&user_bus))
+        tool_oom ("initializing");
+
+      /* If we have an XDG_RUNTIME_DIR and it contains a suitable socket,
+       * dbus-launch --autolaunch can use it, since --autolaunch implies
+       * "I'm OK with getting a bus that is already active".
+       *
+       * (However, plain dbus-launch without --autolaunch must not do so,
+       * because that would break lots of regression tests, which often
+       * use dbus-launch instead of the more appropriate dbus-run-session.)
+       *
+       * At this stage, we just save the user bus's address; later on, the
+       * "babysitter" process will be available to advertise the user-bus
+       * on the X11 display and in ~/.dbus/session-bus, for full
+       * backwards compatibility.
+       */
+      if (!_dbus_lookup_user_bus (&user_bus_supported, &user_bus, &error))
+        {
+          fprintf (stderr, "%s\n", error.message);
+          exit (1);
+        }
+      else if (user_bus_supported)
+        {
+          verbose ("=== Using existing user bus \"%s\"\n",
+                   _dbus_string_get_const_data (&user_bus));
+        }
+      else
+        {
+          _dbus_string_free (&user_bus);
         }
 
       verbose ("Autolaunch enabled (using X11).\n");
@@ -1092,6 +1158,22 @@ main (int argc, char **argv)
       close (bus_pid_to_babysitter_pipe[READ_END]);
       close (bus_pid_to_babysitter_pipe[WRITE_END]);
 
+      /* If we have a user bus and want to use it, do so instead of
+       * exec'ing a new dbus-daemon. */
+      if (autolaunch && user_bus_supported)
+        {
+          do_write (bus_pid_to_launcher_pipe[WRITE_END], "0\n", 2);
+          close (bus_pid_to_launcher_pipe[WRITE_END]);
+
+          do_write (bus_address_to_launcher_pipe[WRITE_END],
+                    _dbus_string_get_const_data (&user_bus),
+                    _dbus_string_get_length (&user_bus));
+          do_write (bus_address_to_launcher_pipe[WRITE_END], "\n", 1);
+          close (bus_address_to_launcher_pipe[WRITE_END]);
+
+          exit (0);
+        }
+
       sprintf (write_pid_fd_as_string,
                "%d", bus_pid_to_launcher_pipe[WRITE_END]);
 
@@ -1100,24 +1182,41 @@ main (int argc, char **argv)
 
       verbose ("Calling exec()\n");
  
-#ifdef DBUS_BUILD_TESTS 
-      /* exec from testdir */
-      if (getenv("DBUS_USE_TEST_BINARY") != NULL)
-        {
-          execl (TEST_BUS_BINARY,
-                 TEST_BUS_BINARY,
-                 "--fork",
-                 "--print-pid", write_pid_fd_as_string,
-                 "--print-address", write_address_fd_as_string,
-                 config_file ? "--config-file" : "--session",
-                 config_file, /* has to be last in this varargs list */
-                 NULL); 
+#ifdef DBUS_ENABLE_EMBEDDED_TESTS
+      {
+        const char *test_daemon;
+        /* exec from testdir */
+        if (getenv ("DBUS_USE_TEST_BINARY") != NULL &&
+            (test_daemon = getenv ("DBUS_TEST_DAEMON")) != NULL)
+          {
+            if (config_file == NULL && getenv ("DBUS_TEST_DATA") != NULL)
+              {
+                config_file = concat2 (getenv ("DBUS_TEST_DATA"),
+                    "/valid-config-files/session.conf");
 
-          fprintf (stderr,
-                   "Failed to execute test message bus daemon %s: %s. Will try again with the system path.\n",
-                   TEST_BUS_BINARY, strerror (errno));
-        }
- #endif /* DBUS_BUILD_TESTS */
+                if (config_file == NULL)
+                  {
+                    fprintf (stderr, "Out of memory\n");
+                    exit (1);
+                  }
+              }
+
+            execl (test_daemon,
+                   test_daemon,
+                   "--fork",
+                   "--print-pid", write_pid_fd_as_string,
+                   "--print-address", write_address_fd_as_string,
+                   config_file ? "--config-file" : "--session",
+                   config_file, /* has to be last in this varargs list */
+                   NULL);
+
+            fprintf (stderr,
+                     "Failed to execute test message bus daemon %s: %s.\n",
+                     test_daemon, strerror (errno));
+            exit (1);
+          }
+      }
+ #endif /* DBUS_ENABLE_EMBEDDED_TESTS */
 
       execl (DBUS_DAEMONDIR"/dbus-daemon",
              DBUS_DAEMONDIR"/dbus-daemon",
@@ -1162,7 +1261,6 @@ main (int argc, char **argv)
       char *end;
       long wid = 0;
       long val;
-      int ret2;
 
       verbose ("=== Parent dbus-launch continues\n");
       
@@ -1231,11 +1329,17 @@ main (int argc, char **argv)
 
       bus_pid = val;
 
+      /* Have to initialize bus_pid_to_kill ASAP, so that the
+         X error callback can kill it if an error happens. */
+      bus_pid_to_kill = bus_pid;
+
       close (bus_pid_to_launcher_pipe[READ_END]);
 
 #ifdef DBUS_ENABLE_X11_AUTOLAUNCH
       if (xdisplay != NULL)
         {
+          int ret2;
+
           verbose("Saving x11 address\n");
           ret2 = x11_save_address (bus_address, bus_pid, &wid);
           /* Only get an existing dbus session when autolaunching */
@@ -1245,7 +1349,6 @@ main (int argc, char **argv)
                 {
                   char *address = NULL;
                   /* another window got added. Return its address */
-                  bus_pid_to_kill = bus_pid;
                   if (x11_get_address (&address, &bus_pid, &wid)
                        && address != NULL)
                     {
